# z: use MemHashString and xxhash.Sum64String